+++ cmd/siac/hostcmd.go
@@ -63,11 +63,19 @@ Available settings:
      minstorageprice:           currency / TB / Month
      minuploadbandwidthprice:   currency / TB
 
+     ephemeralaccountexpiry:     seconds
+     maxephemeralaccountbalance: currency
+     maxephemeralaccountrisk:    currency
+
 Currency units can be specified, e.g. 10SC; run 'siac help wallet' for details.
 
 Durations (maxduration and windowsize) must be specified in either blocks (b),
 hours (h), days (d), or weeks (w). A block is approximately 10 minutes, so one
 hour is six blocks, a day is 144 blocks, and a week is 1008 blocks.
+
+Timeouts (ephemeralaccountexpiry) must be specified in either seconds (s),
+hours (h), days (d), or weeks (w). One hour is 3600 seconds, a day is 86400
+seconds, and a week is 604800 seconds.
 
 For a description of each parameter, see doc/API.md.
 
@@ -218,6 +226,10 @@ Host Internal Settings:
 	minstorageprice:           %v / TB / Month
 	minuploadbandwidthprice:   %v / TB
 
+	ephemeralaccountexpiry:     %v
+	maxephemeralaccountbalance: %v
+	maxephemeralaccountrisk:    %v
+
 Host Financials:
 	Contract Count:               %v
 	Transaction Fee Compensation: %v
@@ -261,6 +273,10 @@ RPC Stats:
 			currencyUnits(is.MinDownloadBandwidthPrice.Mul(modules.BytesPerTerabyte)),
 			currencyUnits(is.MinStoragePrice.Mul(modules.BlockBytesPerMonthTerabyte)),
 			currencyUnits(is.MinUploadBandwidthPrice.Mul(modules.BytesPerTerabyte)),
+
+			is.EphemeralAccountExpiry,
+			currencyUnits(is.MaxEphemeralAccountBalance),
+			currencyUnits(is.MaxEphemeralAccountRisk),
 
 			fm.ContractCount, currencyUnits(fm.ContractCompensation),
 			currencyUnits(fm.PotentialContractCompensation),
@@ -340,7 +356,7 @@ func hostconfigcmd(param, value string) {
 	var err error
 	switch param {
 	// currency (convert to hastings)
-	case "collateralbudget", "maxcollateral", "minbaserpcprice", "mincontractprice", "minsectoraccessprice":
+	case "collateralbudget", "maxcollateral", "minbaserpcprice", "mincontractprice", "minsectoraccessprice", "maxephemeralaccountbalance", "maxephemeralaccountrisk":
 		value, err = parseCurrency(value)
 		if err != nil {
 			die("Could not parse "+param+":", err)
@@ -378,6 +394,13 @@ func hostconfigcmd(param, value string) {
 	// duration (convert to blocks)
 	case "maxduration", "windowsize":
 		value, err = parsePeriod(value)
+		if err != nil {
+			die("Could not parse "+param+":", err)
+		}
+
+	// timeout (convert to seconds)
+	case "ephemeralaccountexpiry":
+		value, err = parseTimeout(value)
 		if err != nil {
 			die("Could not parse "+param+":", err)
 		}