+++ node/api/client/host.go
@@ -52,6 +52,15 @@ const (
 	HostParamMaxReviseBatchSize = HostParam("maxrevisebatchsize")
 	// HostParamNetAddress is the announced netaddress of the host.
 	HostParamNetAddress = HostParam("netaddress")
+	// HostParamEphemeralAccountExpiry is the maximum amount of time an
+	// ephemeral account can be inactive before it expires and gets deleted.
+	HostParamEphemeralAccountExpiry = HostParam("ephemeralaccountexpiry")
+	// HostParamMaxEphemeralAccountBalance is the maximum ephemeral account
+	// balance in hastings
+	HostParamMaxEphemeralAccountBalance = HostParam("maxephemeralaccountbalance")
+	// HostParamMaxEphemeralAccountRisk is the maximum ephemeral account risk in
+	// hastings
+	HostParamMaxEphemeralAccountRisk = HostParam("maxephemeralaccountrisk")
 )
 
 // HostAnnouncePost uses the /host/announce endpoint to announce the host to