+++ node/api/host.go
@@ -227,6 +227,30 @@ func (api *API) parseHostSettings(req *http.Request) (modules.HostInternalSettin
 		}
 		settings.MinUploadBandwidthPrice = x
 	}
+	if req.FormValue("ephemeralaccountexpiry") != "" {
+		var x uint64
+		_, err := fmt.Sscan(req.FormValue("ephemeralaccountexpiry"), &x)
+		if err != nil {
+			return modules.HostInternalSettings{}, err
+		}
+		settings.EphemeralAccountExpiry = x
+	}
+	if req.FormValue("maxephemeralaccountbalance") != "" {
+		var x types.Currency
+		_, err := fmt.Sscan(req.FormValue("maxephemeralaccountbalance"), &x)
+		if err != nil {
+			return modules.HostInternalSettings{}, err
+		}
+		settings.MaxEphemeralAccountBalance = x
+	}
+	if req.FormValue("maxephemeralaccountrisk") != "" {
+		var x types.Currency
+		_, err := fmt.Sscan(req.FormValue("maxephemeralaccountrisk"), &x)
+		if err != nil {
+			return modules.HostInternalSettings{}, err
+		}
+		settings.MaxEphemeralAccountRisk = x
+	}
 
 	return settings, nil
 }