+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Latest:
  - FUSE folders can now be mounted with 'AllowOther' as an option
  - Added alerts for when contracts can't be renewed or refreshed
  - Smarter fund allocation when initially forming contracts
+ - When repairing files from disk, an integrity check is performed to ensure
+   that corrupted / altered data is not used to perform repairs
 
 **Bugs Fixed**
  - Repair operations would sometimes perform useless and redundant repairs